engine/requirements.txt
# =============================================================================
# ONTOLOGY ENGINE - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Azure Services
# -----------------------------------------------------------------------------
azure-cosmos>=4.5.0          # Cosmos DB client for telemetry storage

# -----------------------------------------------------------------------------
# Environment & Configuration
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0         # Load environment variables from .env

# -----------------------------------------------------------------------------
# AI / LLM Integration (Optional)
# -----------------------------------------------------------------------------
openai>=1.0.0                # Azure OpenAI for answer scoring

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pyyaml>=6.0                  # Parse YAML lexicon files
numpy>=1.24.0                # Numerical operations for performance stats
